--- PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-hdinsight-base/src/main/java/com/microsoft/azure/cosmosspark/common/Validatable.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- */
-
-package com.microsoft.azure.cosmosspark.common;
-
-import com.microsoft.azuretools.azurecommons.helpers.Nullable;
-
-public interface Validatable {
-    boolean isLegal();
-
-    @Nullable
-    String getErrorMessage();
-}
