+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-sparkoncosmos/build.gradle
@@ -9,9 +9,12 @@ intellij {
 
 dependencies {
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.9.10")
-    implementation project(path: ':azure-intellij-plugin-lib', configuration: 'instrumentedJar')
-    implementation project(path: ':azure-intellij-plugin-lib-java', configuration: 'instrumentedJar')
-    implementation project(path: ':azure-intellij-plugin-hdinsight', configuration: 'instrumentedJar')
+    compileOnly project(path: ':azure-intellij-plugin-lib')
+    runtimeOnly project(path: ':azure-intellij-plugin-lib', configuration: 'instrumentedJar')
+    compileOnly project(path: ':azure-intellij-plugin-lib-java')
+    runtimeOnly project(path: ':azure-intellij-plugin-lib-java', configuration: 'instrumentedJar')
+    compileOnly project(path: ':azure-intellij-plugin-hdinsight')
+    runtimeOnly project(path: ':azure-intellij-plugin-hdinsight', configuration: 'instrumentedJar')
     implementation 'com.microsoft.azure:azure-toolkit-ide-common-lib'
     implementation 'com.microsoft.hdinsight:azure-toolkit-ide-cosmos-spark-lib'
     implementation 'org.dom4j:dom4j:2.1.3'
