+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-sparkoncosmos/src/main/kotlin/com/microsoft/azure/cosmos/spark/actions/SparkAppSubmitContext.kt
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.microsoft.azure.hdinsight.spark.actions
+package com.microsoft.azure.cosmos.spark.actions
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.actionSystem.DataKey
