+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-synapse/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id "java"
+    id "org.jetbrains.kotlin.jvm"
 }
 
 intellij {
@@ -7,6 +8,7 @@ intellij {
 }
 
 dependencies {
+    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.9.10")
     compileOnly project(path: ':azure-intellij-plugin-lib')
     runtimeOnly project(path: ':azure-intellij-plugin-lib', configuration: 'instrumentedJar')
     compileOnly project(path: ':azure-intellij-plugin-lib-java')
