+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-synapse/src/main/resources/META-INF/azure-intellij-plugin-synapse.xml
@@ -3,4 +3,10 @@
         <explorerNodeProvider implementation="com.microsoft.azure.toolkit.ide.synapse.spark.SynapseNodeProvider"/>
         <actions implementation="com.microsoft.azure.toolkit.ide.synapse.spark.SynapseActionsContributor"/>
     </extensions>
+    <actions>
+        <action id="Actions.SubmitArcadiaSparkApplicationAction"
+                class="com.microsoft.azure.synapse.spark.actions.ArcadiaSparkSelectAndSubmitAction"
+                text="Apache Spark on Azure Synapse"
+                description="Submit Apache Spark on Azure Synapse Application" icon="/icons/Spark.png"/>
+    </actions>
 </idea-plugin>
